requirements.txt
gradio>=4.50.0
# reportlab==4.0.7
python-dotenv==1.0.0
jinja2>=3.0.0
openai>=1.62.0
